Reduce unneeded accuracy

+++ synth/src/clock.rs
@@ -22,8 +22,8 @@ impl Clock {
     pub fn t(&self, frequency: f32) -> f32 {
         // I don't believe this works for timers < 1 Hz. This requires some
         // investigation.
-        ((self.time % self.sample_rate) as f64 / self.sample_rate as f64
-            * frequency as f64
-            % 1.) as f32
+        (self.time % self.sample_rate) as f32 / self.sample_rate as f32
+            * frequency
+            % 1.
     }
 }
